FAI-1104 add some validation

+++ src/SFA.DAS.FAA.Web/Controllers/UserController.cs
@@ -79,7 +79,7 @@ namespace SFA.DAS.FAA.Web.Controllers
                 {
                     GovIdentifier = User.Claims.GovIdentifier(),
                     Email = User.Claims.Email(),
-                    DateOfBirth = new DateOnly(model.Year, model.Month, model.Day)
+                    DateOfBirth = new DateOnly((int)model.Year, (int)model.Month, (int)model.Day)
                 };
                 await mediator.Send(command);
             }
